Use GetSearchHintsAsync to search for artists
The ItemQuery api call does not return artists with no albums while a
SearchHints call gets both artists and album artists. I tried to use
GetSearchHintsAsync for everything but couldn't get it to work properly
for albums (art was always missing) so this is what I got to make it
work.